+++ src/App.tsx
@@ -1,4 +1,5 @@
 import { Login } from './components/Login'
+import { ChangelogWidget } from './components/ChangelogWidget'
 
 function App() {
   return (
@@ -13,6 +14,8 @@ function App() {
       {/* Floating Particles (Simulating dust/magic) - Optional Visual Flourish */}
       <div className="absolute top-1/4 left-1/4 w-2 h-2 bg-yellow-400 rounded-full blur-[1px] animate-ping opacity-20"></div>
       <div className="absolute bottom-1/3 right-1/4 w-3 h-3 bg-amber-500 rounded-full blur-[2px] animate-ping opacity-20 delay-1000"></div>
+
+      <ChangelogWidget />
     </div>
   )
 }